Bug#56521 Assertion failed: (m_state == 2), function allocated_to_free, pfs_lock.h (138)
Before this fix, it was possible to build the server:
- with the performance schema
- with a dummy implementation of my_atomic (MY_ATOMIC_MODE_DUMMY).

In this case, the resulting binary will just crash,
as this configuration is not supported.

This fix enforces that the build will fail with a compilation error in this
configuration, instead of resulting in a broken binary.

--BZR--
revision-id: marc.alff@oracle.com-20100909202847-p2eq4vqq8v1r27ir
property-branch-nick: mysql-5.5-bugfixing-56521
testament3-sha1: 44c60a746f20239adfc3f331b8f8262f7d40a566