Bug #54606 innodb fast alter table + pack_keys=0 prevents            adding new indexes
A fast alter table requires that the existing (old) table
and indices are unchanged (i.e only new indices can be
added).  To verify this, the layout and flags of the old
table/indices are compared for equality with the new.

The PACK_KEYS option is a no-op in InnoDB, but the flag
exists, and is used in the table compare.  We need to
check this (table) option flag before deciding whether an 
index should be packed or not.  If the table has
explicitly set PACK_KEYS to 0, the created indices should
not be marked as packed/packable. 

--BZR--
revision-id: magne.mahre@sun.com-20100916105108-f9h90u85ejlonfxz
property-branch-nick: mysql-5.1-bugteam-54606
testament3-sha1: 8e3d191d94650a31ef2551189e76290aad693b37