Bug #54920 Stored functions are allowed in HANDLER statements,            but broken.
Before this patch, it was allowed to use stored functions in
HANDLER ... READ statements. The problem was that this functionality
was not really supported by the code. Proper locking would for example
not be performed, and it was also possible to break replication by
having stored functions that performed updates.

This patch disallows the use of stored functions in HANDLER ... READ.
Any such statement will now give an ER_NOT_SUPPORTED_YET error.
This is an incompatible change and should be reflected in the
documentation.

Test case added to handler_myisam/handler_innodb.test.

--BZR--
revision-id: jon.hauglid@oracle.com-20100924071816-fmk8i6ost9fp170w
property-branch-nick: mysql-5.5-runtime-bug54920
testament3-sha1: 3dddfec76fd5ef804dbbba18392ad9dbd567d15a