Bug#56822: Add a thread state for sessions waiting on the query cache lock
The problem was that threads waiting on the query cache lock
are not easily seen due to the lack of a state indicating that
the thread is waiting on the said lock. This made it difficult
for users to quickly spot (for example, via SHOW PROCESSLIST)
a query cache contention problem.

The solution is to update the thread state when the query cache
lock needs to be acquired. Whenever the lock is to be acquired,
the thread state is updated to "Waiting for query cache lock"
and is reset once the lock is granted or the wait is interrupted.
The intention is to make query cache related hangs more evident.

To further investigate query cache related locking problems, one
may use PERFORMANCE_SCHEMA to track the overhead associated with
the locking bits and determine which particular lock is being a
contention point.
--BZR--
revision-id: davi.arnaut@oracle.com-20101007225137-yq4smmwrxdn4g03z
property-branch-nick: 56822-5.5
property-file-info: ld7:file_id65:sp1f-sql_cache.cc-19700101030959-74bsqwcnhboovijsogcenqana5inu6wo7:message149:Set and reset the thread state whenever a attempt to lock the
property-file-info: query cache is made.
property-file-info: 
property-file-info: Use DEBUG_SYNC instead of the now unnecessary wait_for_kill hack.4:path16:sql/sql_cache.ccee
testament3-sha1: fc2c7f53cd4b4387d63ab26c84e9558aafb0de7d