Fix Bug #54538 - use of exclusive innodb dictionary lock limits performance.
This patch doesn't get rid of the need to acquire the dict_sys->mutex but
reduces the need to keep the mutex locked for the duration of the query
to fsp_get_available_space_in_free_extents() from ha_innobase::info().
 
This is a port of revno:3548 from the builtin to the plugin.

rb://501 Approved by Jimmy Yang and Marko Makela.

--BZR--
revision-id: sunny.bains@oracle.com-20101103014053-rlghb3z1wfbc45al
property-branch-nick: 5.1
testament3-sha1: 33adfbb1e1029b37441a2abf6fec08ecfaaf9ce0