Bug#57609 performance_schema does not work with lower_case_table_names
Before this fix, the performance schema tables were defined in UPPERCASE.

This was incompatible with the lowercase_table_names option, and caused
issues with the install / upgrade process, when changing the lower case
table names setting *after* the install or upgrade.

With this fix, all performance schema tables are exposed with lowercase names.
As a result, the name of the performance schema table is always lowercase,
no matter how / if / when the lowercase_table_names setting if changed.

--BZR--
revision-id: marc.alff@oracle.com-20101103154233-niuph86t5rgow1ac
property-branch-nick: mysql-5.5-bugteam-57609
testament3-sha1: b609a65542906db5858dc563594b33f2d58b6cd1