Bug#47350 Support innodb plugin without separate shared object
Add a _commented_ workaround for Bug#47350. The full solution is tricky
to get right as explained in the bug report. It is not worth the effort
to extend the deprecated autotools framework to support conflicting
plugins and would be too risky for MySQL 5.1 (GA).

--BZR--
revision-id: vasil.dimov@oracle.com-20101125065547-1ywiiom8ulvgwyww
property-branch-nick: mysql-5.1-innodb
testament3-sha1: 5e2eee5ff2dd27b1bb5de275fd38102acffca310