Bug#58621 perfschema.dml_ews_by_thread_by_event_name test failure
This fix affects the test suite only.

Before this fix, performance schema tests dml_*.test could
fail with spurious failure, depending on the table content.

This fix simplifies the SELECT tests in the dml_*.test scripts,
to only verify that the SELECT operation passed the security checks
and succeeded, which was the original intent of the test.

Usage of 
  --replace_column 1 # 2 # 3 # 4 # ...
to discard the test output was replaced by a simpler and more maintainable 
  --disable_result_log
which also work for empty tables.

--BZR--
revision-id: marc.alff@oracle.com-20101201080104-j851s2tsdj1rarr6
property-branch-nick: mysql-5.5-bugteam-58621
testament3-sha1: 5443063af72374ad991adc449c62516f1736139d