BUG#46697: Table name in error message is not populated
When a query fails with a different error on the slave,
the sql thread outputs a message (M) containing:

  1. the error message format for the master error code
  2. the master error code
  3. the error message for the slave's error code
  4. the slave error code

Given that the slave has no information on the error message
itself that the master outputs, it can only print its own
version of the message format (but stripped from the 
additional data if the message format requires). This may
confuse users.

To fix this we augment the slave's message (M) to explicitly
state that the master's message is actually an error message 
format, the one associated with the given master error code 
and that the slave server knows about.
--BZR--
revision-id: luis.soares@oracle.com-20101203171745-vdi4h46214c41uyl
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: faf1632b70c1aa459fad23b9343dbcbda64dbd2c