BUG#59267 LOAD DATA LOCAL INFILE not executed on slave with SBR
On windows, an #endif in a wrong place was causing an early
return from mysql_load and thus the LOAD DATA LOCAL was not
executed. This problem was fixed by moving the #endif to the
right place.

The following code was missing
    
  if ((stat_info.st_mode & S_IFIFO) == S_IFIFO)
    is_fifo = 1;

which is required to properly configure and read from the
IO_CACHE when a named pipe is used. So it was re-introduced
before the #endif.
--BZR--
revision-id: alfranio.correia@oracle.com-20110112093132-a64cp6jx61dl0vp3
property-branch-nick: mysql-5.5-security
testament3-sha1: 39ae3cd93fc6892963d7a77ab80a67c9b0ec9783