Bug #59453: Non-ASCIIZ string kills protocol extensibility in MySQL 5.5
When the server sends the name of the plugin it's using in the handshake 
packet it was null terminating it in it's buffer, but was sending a length of 
the packet 1 byte short.
Fixed to send the terminating 0 as well by increasing the length of the 
packet to include it. 
In this way the handshake packet becomes similar to the change user packet
where the plugin name is null terminated.
No test suite added as the fix can only be observed by analyzing the bytes
sent over the wire.
--BZR--
revision-id: georgi.kodinov@oracle.com-20110114154811-9uac3b7a9ra5ipv3
property-branch-nick: B59453-5.5
property-rebase-of: georgi.kodinov@oracle.com-20110114154811-1ni3az7syb027wk1
testament3-sha1: f1a3246ac1b9560c40f4df9f2d20dd7a159a7d03